# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "DriverInputResource" : {
        "properties" : {
          "email" : {
            "maxLength" : 160,
            "minLength" : 0,
            "type" : "string"
          },
          "firstName" : {
            "maxLength" : 80,
            "minLength" : 0,
            "type" : "string"
          },
          "lastName" : {
            "maxLength" : 80,
            "minLength" : 0,
            "type" : "string"
          },
          "licenseNumber" : {
            "maxLength" : 60,
            "minLength" : 0,
            "type" : "string"
          },
          "phoneNumber" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "status" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "email", "firstName", "lastName", "licenseNumber", "phoneNumber" ],
        "type" : "object"
      },
      "DriverV2Resource" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastName" : {
            "type" : "string"
          },
          "licenseNumber" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "EligibilityResource" : {
        "properties" : {
          "outcome" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TankerV2Resource" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "brand" : {
            "type" : "string"
          },
          "capacity" : {
            "format" : "double",
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "licensePlate" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@primefuel.com",
      "name" : "Soporte de FullTank",
      "url" : "https://primefuel.com/support"
    },
    "description" : "API de FullTank para gestionar abastecimiento, flota, entregas, inventario y pagos por organizaciÃ³n.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FullTank Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v2/drivers" : {
      "get" : {
        "description" : "Devuelve los conductores registrados para el tenant distribuidor autenticado.",
        "operationId" : "list_2",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DriverV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de conductores."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DriverV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          }
        },
        "summary" : "Listar conductores",
        "tags" : [ "Conductores de flota" ]
      },
      "post" : {
        "description" : "Crea un conductor para el tenant distribuidor autenticado, identificado a partir del principal.",
        "operationId" : "register_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DriverInputResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Conductor registrado."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El cuerpo no supera la validación o los datos del conductor no son válidos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          }
        },
        "summary" : "Registrar conductor",
        "tags" : [ "Conductores de flota" ]
      }
    },
    "/api/v2/drivers/eligible" : {
      "get" : {
        "description" : "Devuelve conductores activos del tenant autenticado que pueden asignarse y no están ocupados.",
        "operationId" : "listEligible_1",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DriverV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de conductores elegibles."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DriverV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          }
        },
        "summary" : "Listar conductores elegibles",
        "tags" : [ "Conductores de flota" ]
      }
    },
    "/api/v2/drivers/{driverId}" : {
      "get" : {
        "description" : "Devuelve el conductor indicado si pertenece al tenant distribuidor autenticado.",
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverV2Resource"
                }
              }
            },
            "description" : "Conductor devuelto."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverV2Resource"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverV2Resource"
                }
              }
            },
            "description" : "El conductor no existe o pertenece a otro tenant distribuidor."
          }
        },
        "summary" : "Consultar conductor por identificador",
        "tags" : [ "Conductores de flota" ]
      },
      "put" : {
        "description" : "Aplica los cambios recibidos a un conductor del tenant distribuidor autenticado.",
        "operationId" : "update_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DriverInputResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Conductor actualizado."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El cuerpo no supera la validación o los datos del conductor no son válidos."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El conductor no existe o no pertenece al tenant autenticado."
          }
        },
        "summary" : "Actualizar conductor",
        "tags" : [ "Conductores de flota" ]
      }
    },
    "/api/v2/drivers/{driverId}/activate" : {
      "post" : {
        "description" : "Reactiva un conductor desactivado que pertenece al tenant distribuidor autenticado.",
        "operationId" : "activate_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Conductor reactivado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El conductor no existe o no pertenece al tenant autenticado."
          }
        },
        "summary" : "Reactivar conductor",
        "tags" : [ "Conductores de flota" ]
      }
    },
    "/api/v2/drivers/{driverId}/deactivate" : {
      "post" : {
        "description" : "Desactiva el conductor del tenant autenticado sin eliminar su registro.",
        "operationId" : "deactivate_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Conductor desactivado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El conductor no existe o no pertenece al tenant autenticado."
          }
        },
        "summary" : "Desactivar conductor",
        "tags" : [ "Conductores de flota" ]
      }
    },
    "/api/v2/drivers/{driverId}/eligibility" : {
      "get" : {
        "description" : "Devuelve el resultado y motivo de elegibilidad para un conductor del tenant autenticado.",
        "operationId" : "eligibility_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "Evaluación de elegibilidad devuelta."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "El conductor no existe o pertenece a otro tenant distribuidor."
          }
        },
        "summary" : "Evaluar elegibilidad del conductor",
        "tags" : [ "Conductores de flota" ]
      }
    },
...
    "/api/v2/tankers/eligible" : {
      "get" : {
        "description" : "Devuelve cisternas activas del tenant autenticado que pueden asignarse y no están ocupadas.",
        "operationId" : "listEligible",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TankerV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de cisternas elegibles."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TankerV2Resource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          }
        },
        "summary" : "Listar cisternas elegibles",
        "tags" : [ "Cisternas de flota" ]
      }
    },
...
    "/api/v2/tankers/{tankerId}/activate" : {
      "post" : {
        "description" : "Reactiva una cisterna desactivada que pertenece al tenant distribuidor autenticado.",
        "operationId" : "activate",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Cisterna reactivada."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La cisterna no existe o no pertenece al tenant autenticado."
          }
        },
        "summary" : "Reactivar cisterna",
        "tags" : [ "Cisternas de flota" ]
      }
    },
    "/api/v2/tankers/{tankerId}/deactivate" : {
      "post" : {
        "description" : "Desactiva una cisterna del tenant autenticado sin eliminar su registro.",
        "operationId" : "deactivate",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Cisterna desactivada."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La cisterna no existe o no pertenece al tenant autenticado."
          }
        },
        "summary" : "Desactivar cisterna",
        "tags" : [ "Cisternas de flota" ]
      }
    },
    "/api/v2/tankers/{tankerId}/eligibility" : {
      "get" : {
        "description" : "Devuelve el resultado y motivo de elegibilidad para una cisterna del tenant autenticado.",
        "operationId" : "eligibility",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "Evaluación de elegibilidad devuelta."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "El usuario autenticado no tiene identidad de distribuidor."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResource"
                }
              }
            },
            "description" : "La cisterna no existe o pertenece a otro tenant distribuidor."
          }
        },
        "summary" : "Evaluar elegibilidad de cisterna",
        "tags" : [ "Cisternas de flota" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}